Cleanup, watchdir, multiprocessing
Small code cleanup was done in the configuration file and worker.

The application now also calculates the imphash
(https://www.mandiant.com/blog/tracking-malware-import-hashing/)

Renamed the application to MalZoo

Added directory monitoring (watchdog) with the -w argument

Rewritten the multiprocessing part of the code so it works more
efficient and integrates well with the directory monitor.